+++ ejIntrn.h
@@ -180,7 +180,7 @@ typedef struct ej {
 	int			tid;							/* Current token id */
 	int			eid;							/* Halloc handle */
 	int			flags;							/* Flags */
-	int			userHandle;						/* User defined handle */
+	long		userHandle;						/* User defined handle */
 } ej_t;
 
 /******************************** Prototypes **********************************/
@@ -196,7 +196,7 @@ extern void		*ejGetGlobalFunction(int eid, char_t *name);
 extern int 		ejSetGlobalFunctionDirect(sym_fd_t functions, char_t *name, 
 					int (*fn)(int eid, void *handle, int argc, char_t **argv));
 extern void 	ejError(ej_t* ep, char_t* fmt, ...);
-extern void		ejSetUserHandle(int eid, int handle);
+extern void		ejSetUserHandle(int eid, long handle);
 extern int		ejGetUserHandle(int eid);
 extern int		ejGetLineNumber(int eid);
 extern char_t	*ejGetResult(int eid);
